Split out Key Generation
