+++ lib/generators/random.js
