+++ lib/proxy_signer.js
@@ -1,4 +1,4 @@
-var Random = require('./random');
+var Random = require('./generators/random');
 var Transaction = require('ethereumjs-tx');
 var util = require("ethereumjs-util");
 var txutils = require('./txutils');
