+++ lib/vault.js
@@ -1,5 +1,5 @@
 var nacl = require('tweetnacl');
-var Random = require('./random');
+var Random = require('./generators/random');
 var scrypt = require('scrypt-async');
 
 // TODO come up with a good way to serialize
